Changed the way isPositiveAndBelow is written to avoid needing to cast the second parameter to an int

+++ examples/audio plugin host/Source/FilterIOConfiguration.cpp
@@ -312,7 +312,7 @@ private:
 
                     if (selectedNumChannels != bus->getLastEnabledLayout().size())
                     {
-                        if (isPositiveAndBelow (selectedNumChannels, (int) AudioChannelSet::maxChannelsOfNamedLayout)
+                        if (isPositiveAndBelow (selectedNumChannels, AudioChannelSet::maxChannelsOfNamedLayout)
                              && bus->setCurrentLayoutWithoutEnabling (bus->supportedLayoutWithChannels (selectedNumChannels)))
                         {
                             if (auto* config = owner.getConfig (! isInput))
